Test implementation of Bitmap algorithm

+++ src/load_char.h
@@ -5,9 +5,11 @@ class CharMaps{
     public:
     int map_len;
     char* chars;
-    int* lumin;
-    int* lumin_non_empty;
+    
+    float* lumin;
+    float* lumin_non_empty;
     int lumin_len;
+
     int*** char_bitmaps;
 
     CharMaps();
